refactored apexgrammar class, its method create had been splitted into new classes
